# [None][perf] Disable Swap AB when num tokens exceeds N dimension