--- a/cpp/tensorrt_llm/kernels/cutlass_kernels/moe_gemm/moe_gemm_template_dispatch.h
+++ b/cpp/tensorrt_llm/kernels/cutlass_kernels/moe_gemm/moe_gemm_template_dispatch.h
@@ -577,7 +577,7 @@ MoeGemmRunner<T, WeightType, OutputType, ScaleBiasType>::getTmaWarpSpecializedCo
             return config;
         });
 
-    if (!use_w4_groupwise)
+    if (use_w4_groupwise)
     {
         // w4 groupwise implementation requires swap_ab to be true
         tma_ws_configs.erase(