<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **New Features**
  * Optional A/B swap (swap_ab) added end-to-end: configs, runtime flag, kernel variants, generation tooling, and observable kernel names.
  * SM90 epilogue: per‑column bias and new per‑row/column scale+bias scatter modes.

* **Refactor**
  * Public A/B terminology renamed to Activation/Weight across workspace, pointers, strides, scaling, diagnostics, and string output.
  * Launcher/profiler/workspace caching/sizing extended to handle fusion × swap_ab combos; added transposed final‑output stride and expanded workspace inputs; launcher dispatch updated to encode swap_ab.

* **Other**
  * Min‑latency path removed; test diagnostic strings and version manifests updated.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!--
Please write the PR title by following this template:

**[JIRA ticket/NVBugs ID/GitHub issue/None][type] Summary**

Valid ticket formats:
  - JIRA ticket: [TRTLLM-1234] or [FOOBAR-123] for other FOOBAR project
  - NVBugs ID: [https://nvbugs/1234567]
  - GitHub issue: [#1234]
  - No ticket: [None]

Valid types (lowercase): [fix], [feat], [doc], [infra], [chore], etc.

Examples:
  - [TRTLLM-1234][feat] Add new feature
  - [https://nvbugs/1234567][fix] Fix some bugs
  - [#1234][doc] Update documentation
  - [None][chore] Minor clean-up

Alternative (faster) way using CodeRabbit AI:

**[JIRA ticket/NVBugs ID/GitHub issue/None] @coderabbitai title**

NOTE: "@coderabbitai title" will be replaced by the title generated by CodeRabbit AI, that includes the "[type]" and title.
For more info, see /.coderabbit.yaml.

-->

## Description

<!--
Please explain the issue and the solution in short.
-->

Perf benefits

GB200 DeepSeek 4GPUs, we see a decrease in average TTFT: 3652.8973->3563.9274 (2.5%)

Optimized:  

```
===========================================================
= PERFORMANCE OVERVIEW 
===========================================================
Request Throughput (req/sec):                     7.3411
Total Output Throughput (tokens/sec):             7517.2584
Total Token Throughput (tokens/sec):              15034.5167
Total Latency (ms):                               836934.9168
Average request latency (ms):                     138784.6399
Per User Output Throughput [w/ ctx] (tps/user):   7.3984
Per GPU Output Throughput (tps/gpu):              1879.3146
Average time-to-first-token [TTFT] (ms):          3563.9274
Average time-per-output-token [TPOT] (ms):        132.1806
Per User Output Speed (tps/user):                 7.5710
```

Main branch  
```
===========================================================
= PERFORMANCE OVERVIEW 
===========================================================
Request Throughput (req/sec):                     7.3213
Total Output Throughput (tokens/sec):             7497.0163
Total Token Throughput (tokens/sec):              14994.0326
Total Latency (ms):                               839194.6528
Average request latency (ms):                     139167.8365
Per User Output Throughput [w/ ctx] (tps/user):   7.3780
Per GPU Output Throughput (tps/gpu):              1874.2541
Average time-to-first-token [TTFT] (ms):          3652.8973
Average time-per-output-token [TPOT] (ms):        132.4682
Per User Output Speed (tps/user):                 7.5549
```


<details>
Script:   

```
max_batch_size=1024
max_num_tokens=2176
kv_fraction=0.7
multi_round=6 # change this to 1 to make it faster
num_prompts=$((${max_batch_size} * ${multi_round}))


cat <<EOF > /tmp/extra-llm-api-config.yml
cuda_graph_config:
 enable_padding: true
 batch_sizes: [1, 2, 4, 8, 16, 24, 32, 40, 48, 56, 64, 72, 80, 88, 96, 104, 112, 120, 128]
moe_config:
    backend: WIDEEP
print_iter_log: true
kv_cache_config:
    dtype: fp8
enable_attention_dp: true
EOF


python benchmarks/cpp/prepare_dataset.py \
--tokenizer=${model_path} --stdout token-norm-dist --num-requests=${num_prompts} \
       --input-mean=1024 --output-mean=1024 --input-stdev=0 --output-stdev=0 > /tmp/dataset.json



NUM_GPUS=4

cmd="${cmd} \
mpirun -np ${NUM_GPUS} --allow-run-as-root --oversubscribe trtllm-bench -m ${model_card} --model_path ${model_path} throughput \
       --tp ${NUM_GPUS} \
       --ep ${NUM_GPUS} \
       --warmup 1 \
       --dataset ${dataset_path} \
       --backend pytorch \
       --max_batch_size ${max_batch_size} \
       --max_num_tokens ${max_num_tokens} \
       --kv_cache_free_gpu_mem_fraction ${kv_fraction} \
       --concurrency ${max_batch_size} \
       --extra_llm_api_options /tmp/extra-llm-api-config.yml \
       --num_requests ${num_prompts} \
       --streaming"

eval ${cmd}
```

Optimized:  

```
===========================================================
= PYTORCH BACKEND
===========================================================
Model:   deepseek-ai/DeepSeek-R1
Model Path:  /lustre/fsw/coreai_devtech_all/dastokes/DeepSeek-R1-0528-FP4
TensorRT-LLM Version: 1.1.0rc1
Dtype:   bfloat16
KV Cache Dtype:  FP8
Quantization:  NVFP4

===========================================================
= REQUEST DETAILS 
===========================================================
Number of requests:             6144
Number of concurrent requests:  1018.8281
Average Input Length (tokens):  1024.0000
Average Output Length (tokens): 1024.0000
===========================================================
= WORLD + RUNTIME INFORMATION 
===========================================================
TP Size:                4
PP Size:                1
EP Size:                4
Max Runtime Batch Size: 1024
Max Runtime Tokens:     2176
Scheduling Policy:      GUARANTEED_NO_EVICT
KV Memory Percentage:   70.00%
Issue Rate (req/sec):   1.1684E+17

===========================================================
= PERFORMANCE OVERVIEW 
===========================================================
Request Throughput (req/sec):                     7.3411
Total Output Throughput (tokens/sec):             7517.2584
Total Token Throughput (tokens/sec):              15034.5167
Total Latency (ms):                               836934.9168
Average request latency (ms):                     138784.6399
Per User Output Throughput [w/ ctx] (tps/user):   7.3984
Per GPU Output Throughput (tps/gpu):              1879.3146
Average time-to-first-token [TTFT] (ms):          3563.9274
Average time-per-output-token [TPOT] (ms):        132.1806
Per User Output Speed (tps/user):                 7.5710

-- Per-Request Time-per-Output-Token [TPOT] Breakdown (ms)

[TPOT] MINIMUM: 115.6459
[TPOT] MAXIMUM: 134.7338
[TPOT] AVERAGE: 132.1806
[TPOT] P50    : 133.3448
[TPOT] P90    : 134.0174
[TPOT] P95    : 134.1346
[TPOT] P99    : 134.3340

-- Per-Request Time-to-First-Token [TTFT] Breakdown (ms) 

[TTFT] MINIMUM: 397.7384
[TTFT] MAXIMUM: 31514.2690
[TTFT] AVERAGE: 3563.9274
[TTFT] P50    : 915.1050
[TTFT] P90    : 12625.9686
[TTFT] P95    : 21898.8528
[TTFT] P99    : 29764.1760

-- Per-Request Generation Throughput [GTPS] Breakdown (tps/user)

[GTPS] MINIMUM: 7.4220
[GTPS] MAXIMUM: 8.6471
[GTPS] AVERAGE: 7.5710
[GTPS] P50    : 7.4994
[GTPS] P90    : 7.7257
[GTPS] P95    : 8.1238
[GTPS] P99    : 8.5436

-- Request Latency Breakdown (ms) -----------------------

[Latency] P50    : 137869.8585
[Latency] P90    : 147954.3023
[Latency] P95    : 156814.9808
[Latency] P99    : 164179.1244
[Latency] MINIMUM: 119062.6340
[Latency] MAXIMUM: 165975.8334
[Latency] AVERAGE: 138784.6399

===========================================================
= DATASET DETAILS
===========================================================
Dataset Path:         /tmp/dataset.json
Number of Sequences:  6144

-- Percentiles statistics ---------------------------------

        Input              Output           Seq. Length
-----------------------------------------------------------
MIN:  1024.0000          1024.0000          2048.0000
MAX:  1024.0000          1024.0000          2048.0000
AVG:  1024.0000          1024.0000          2048.0000
P50:  1024.0000          1024.0000          2048.0000
P90:  1024.0000          1024.0000          2048.0000
P95:  1024.0000          1024.0000          2048.0000
P99:  1024.0000          1024.0000          2048.0000
===========================================================
```

Main:   

```
===========================================================
= PYTORCH BACKEND
===========================================================
Model:   deepseek-ai/DeepSeek-R1
Model Path:  /lustre/fsw/coreai_devtech_all/dastokes/DeepSeek-R1-0528-FP4
TensorRT-LLM Version: 1.1.0rc1
Dtype:   bfloat16
KV Cache Dtype:  FP8
Quantization:  NVFP4

===========================================================
= REQUEST DETAILS 
===========================================================
Number of requests:             6144
Number of concurrent requests:  1018.8902
Average Input Length (tokens):  1024.0000
Average Output Length (tokens): 1024.0000
===========================================================
= WORLD + RUNTIME INFORMATION 
===========================================================
TP Size:                4
PP Size:                1
EP Size:                4
Max Runtime Batch Size: 1024
Max Runtime Tokens:     2176
Scheduling Policy:      GUARANTEED_NO_EVICT
KV Memory Percentage:   70.00%
Issue Rate (req/sec):   1.1715E+17

===========================================================
= PERFORMANCE OVERVIEW 
===========================================================
Request Throughput (req/sec):                     7.3213
Total Output Throughput (tokens/sec):             7497.0163
Total Token Throughput (tokens/sec):              14994.0326
Total Latency (ms):                               839194.6528
Average request latency (ms):                     139167.8365
Per User Output Throughput [w/ ctx] (tps/user):   7.3780
Per GPU Output Throughput (tps/gpu):              1874.2541
Average time-to-first-token [TTFT] (ms):          3652.8973
Average time-per-output-token [TPOT] (ms):        132.4682
Per User Output Speed (tps/user):                 7.5549

-- Per-Request Time-per-Output-Token [TPOT] Breakdown (ms)

[TPOT] MINIMUM: 115.7558
[TPOT] MAXIMUM: 135.7859
[TPOT] AVERAGE: 132.4682
[TPOT] P50    : 133.2145
[TPOT] P90    : 135.3390
[TPOT] P95    : 135.3831
[TPOT] P99    : 135.4911

-- Per-Request Time-to-First-Token [TTFT] Breakdown (ms) 

[TTFT] MINIMUM: 330.1804
[TTFT] MAXIMUM: 32081.9844
[TTFT] AVERAGE: 3652.8973
[TTFT] P50    : 1126.0854
[TTFT] P90    : 12834.2514
[TTFT] P95    : 22279.2546
[TTFT] P99    : 30309.4881

-- Per-Request Generation Throughput [GTPS] Breakdown (tps/user)

[GTPS] MINIMUM: 7.3645
[GTPS] MAXIMUM: 8.6389
[GTPS] AVERAGE: 7.5549
[GTPS] P50    : 7.5067
[GTPS] P90    : 7.6858
[GTPS] P95    : 8.0950
[GTPS] P99    : 8.5312

-- Request Latency Breakdown (ms) -----------------------

[Latency] P50    : 138556.0709
[Latency] P90    : 148124.5726
[Latency] P95    : 157174.1736
[Latency] P99    : 164704.1215
[Latency] MINIMUM: 119430.6584
[Latency] MAXIMUM: 166665.2151
[Latency] AVERAGE: 139167.8365

===========================================================
= DATASET DETAILS
===========================================================
Dataset Path:         /tmp/dataset.json
Number of Sequences:  6144

-- Percentiles statistics ---------------------------------

        Input              Output           Seq. Length
-----------------------------------------------------------
MIN:  1024.0000          1024.0000          2048.0000
MAX:  1024.0000          1024.0000          2048.0000
AVG:  1024.0000          1024.0000          2048.0000
P50:  1024.0000          1024.0000          2048.0000
P90:  1024.0000          1024.0000          2048.0000
P95:  1024.0000          1024.0000          2048.0000
P99:  1024.0000          1024.0000          2048.0000
===========================================================
```


</details>

## PR Checklist

Please review the following before submitting your PR:
- PR description clearly explains what and why. If using CodeRabbit's summary, please make sure it makes sense.
- PR Follows [TRT-LLM CODING GUIDELINES](https://github.com/NVIDIA/TensorRT-LLM/blob/main/CODING_GUIDELINES.md) to the best of your knowledge.
- Test cases are provided for new code paths (see [test instructions](https://github.com/NVIDIA/TensorRT-LLM/tree/main/tests#1-how-does-the-ci-work))
- Any new dependencies have been scanned for license and vulnerabilities
- [CODEOWNERS](https://github.com/NVIDIA/TensorRT-LLM/blob/main/.github/CODEOWNERS) updated if ownership changes
- Documentation updated as needed
- The reviewers assigned automatically/manually are appropriate for the PR.


- [x] Please check this after reviewing the above items as appropriate for this PR.


## Test Coverage

<!--
Please list clearly what are the relevant test(s) that can safeguard the changes in the PR. This helps us to ensure we have sufficient test coverage for the PR.
-->

## GitHub Bot Help

`/bot [-h] ['run', 'kill', 'skip', 'reuse-pipeline'] ...`

Provide a user friendly way for developers to interact with a Jenkins server.

Run `/bot [-h|--help]` to print this help message.

See details below for each supported subcommand.

<details>

`run  [--reuse-test (optional)pipeline-id --disable-fail-fast --skip-test --stage-list "A10-PyTorch-1, xxx" --gpu-type "A30, H100_PCIe" --test-backend "pytorch, cpp" --add-multi-gpu-test --only-multi-gpu-test --disable-multi-gpu-test --post-merge --extra-stage "H100_PCIe-TensorRT-Post-Merge-1, xxx" --detailed-log --debug(experimental)]`

Launch build/test pipelines. All previously running jobs will be killed.

`--reuse-test (optional)pipeline-id ` *(OPTIONAL)* : Allow the new pipeline to reuse build artifacts and skip successful test stages from a specified pipeline or the last pipeline if no pipeline-id is indicated. If the Git commit ID has changed, this option will be always ignored. The DEFAULT behavior of the bot is to reuse build artifacts and successful test results from the last pipeline.

`--disable-reuse-test ` *(OPTIONAL)* : Explicitly prevent the pipeline from reusing build artifacts and skipping successful test stages from a previous pipeline. Ensure that all builds and tests are run regardless of previous successes.

`--disable-fail-fast ` *(OPTIONAL)* : Disable fail fast on build/tests/infra failures.

`--skip-test ` *(OPTIONAL)* : Skip all test stages, but still run build stages, package stages and sanity check stages. Note: Does **NOT** update GitHub check status.

`--stage-list "A10-PyTorch-1, xxx"` *(OPTIONAL)* : Only run the specified test stages. Examples: "A10-PyTorch-1, xxx". Note: Does **NOT** update GitHub check status.

`--gpu-type "A30, H100_PCIe"` *(OPTIONAL)* : Only run the test stages on the specified GPU types. Examples: "A30, H100_PCIe". Note: Does **NOT** update GitHub check status.

`--test-backend "pytorch, cpp"` *(OPTIONAL)* : Skip test stages which don't match the specified backends. Only support [pytorch, cpp, tensorrt, triton]. Examples: "pytorch, cpp" (does not run test stages with tensorrt or triton backend). Note: Does **NOT** update GitHub pipeline status.

`--only-multi-gpu-test ` *(OPTIONAL)* : Only run the multi-GPU tests. Note: Does **NOT** update GitHub check status.

`--disable-multi-gpu-test ` *(OPTIONAL)* : Disable the multi-GPU tests. Note: Does **NOT** update GitHub check status.

`--add-multi-gpu-test ` *(OPTIONAL)* : Force run the multi-GPU tests in addition to running L0 pre-merge pipeline.

`--post-merge ` *(OPTIONAL)* : Run the L0 post-merge pipeline instead of the ordinary L0 pre-merge pipeline.

`--extra-stage "H100_PCIe-TensorRT-Post-Merge-1, xxx"` *(OPTIONAL)* : Run the ordinary L0 pre-merge pipeline and specified test stages. Examples: --extra-stage "H100_PCIe-TensorRT-Post-Merge-1, xxx".

`--detailed-log ` *(OPTIONAL)* : Enable flushing out all logs to the Jenkins console. This will significantly increase the log volume and may slow down the job.

`--debug ` *(OPTIONAL)* : **Experimental feature**. Enable access to the CI container for debugging purpose. Note: Specify exactly one stage in the `stage-list` parameter to access the appropriate container environment. Note: Does **NOT** update GitHub check status.

For guidance on mapping tests to stage names, see `docs/source/reference/ci-overview.md`
and the `scripts/test_to_stage_mapping.py` helper.

### kill

`kill  `

Kill all running builds associated with pull request.

### skip

`skip --comment COMMENT `

Skip testing for latest commit on pull request. `--comment "Reason for skipping build/test"` is required. IMPORTANT NOTE: This is dangerous since lack of user care and validation can cause top of tree to break.

### reuse-pipeline

`reuse-pipeline `

Reuse a previous pipeline to validate current commit. This action will also kill all currently running builds associated with the pull request. IMPORTANT NOTE: This is dangerous since lack of user care and validation can cause top of tree to break.

</details>