@djns99 2025-08-22
/bot run --disable-fail-fast