@djns99 2025-08-25
/bot run --disable-fail-fast